+++ eptnr_package/requirements.txt
@@ -3,11 +3,15 @@ geopandas==0.11.1
 numpy==1.23.3
 torch==1.12.1
 matplotlib==3.6.0
-igraph==0.10.1
+igraph==0.9.10
 networkx==2.8.6
 osmnx==1.2.2
 pygad==2.18.1
 tqdm==4.64.1
 pyarrow==9.0.0
+pytest==7.1.3
+ray==2.0.0
 urbanaccess==0.2.2
-jupyterlab==3.4.7
+statsmodels==0.13.2
+inequality==1.0.0
+cairocffi==1.4.0
