+++ notebooks/requirements.txt
@@ -1,0 +1,3 @@
+jupyterlab==3.4.7
+
+-e ../eptnr_package
